Move writable-subject to state-lib folder and create new state lib
SvelteState allows to easily manage interdependent stores, preventing the need for a lot of combineLatest and other forkJoin pipes.
