+++ libs/search-widget/src/core/state-lib/index.ts
@@ -1,0 +1,2 @@
+export * from './state.lib';
+export * from './writable-subject';
